refactor: Reorganize enhanced utilities into proper src/ subdirectories
- Move JavaScript files from lib/ to src/lib/ and src/utils/
- Move MCP resources from lib/mcp/ to src/prompts/ and src/resources/
- Update package.json exports to reflect new file locations
- Update import paths in BaseAPIEnhanced and test files
- Update README documentation with new project structure
- Maintain backward compatibility through package.json exports

New structure:
- src/lib/ - Core library utilities (api-response-utils, base-api-enhanced, etc.)
- src/utils/ - Utility classes (logger, resource-loader)
- src/prompts/ - Framework default prompts
- src/resources/ - Framework default resources

All tests passing with new structure.
