+++ src/lib/wordpress-source-manager.js
