+++ src/prompts/health-service-monitoring.json
