+++ src/prompts/wordpress-content-creation.json
