+++ src/prompts/youtube-video-analysis.json
@@ -1,5 +1,4 @@
 
-ed 
 {
   "name": "youtube-video-analysis",
   "description": "Prompt for analyzing YouTube videos and extracting insights",
