+++ src/resources/health-monitoring-guide.md
