+++ src/resources/wordpress-content-guide.md
